system/intention_recognition/requirements.txt
mediapipe~=0.8.3
opencv-python~=4.5.2.52
pandas~=1.1.5
numpy~=1.19.3
pyrealsense2~=2.47.0.3313
matplotlib~=3.3.4
scikit-image~=0.17.2
Pillow~=8.2.0
rospy~=1.14.12
rosbag~=1.14.12
colorlog~=6.6.0
sktime~=0.8.0
scikit-learn~=0.24.2
scipy~=1.5.4
joblib~=1.0.1
statsmodels~=0.12.2
numba~=0.53.1
pathlib~=1.0.1
setuptools~=56.2.0
python-dotenv~=0.19.2